Move hf_hub_url to hub module
