+++ src/datasets/filesystems/hffilesystem.py
@@ -5,7 +5,8 @@ import fsspec
 from fsspec import AbstractFileSystem
 from huggingface_hub.hf_api import DatasetInfo
 
-from ..utils.file_utils import get_authentication_headers_for_url, hf_hub_url
+from ..utils.file_utils import get_authentication_headers_for_url
+from ..utils.hub import hf_hub_url
 
 
 class HfFileSystem(AbstractFileSystem):
